Refactor out redundant rpc-client ops methods. Implement a first-draft of net_peers business logic. Add net peers and net listeners CLI commands.
