+++ forest/src/cli/auth_cmd.rs
@@ -1,9 +1,12 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use super::print_rpc_res;
-use rpc_client::auth_new;
+use jsonrpc_v2::Error as JsonRpcError;
 use structopt::StructOpt;
+
+use super::{handle_rpc_err, print_rpc_res_bytes};
+use auth::*;
+use rpc_client::auth_ops::*;
 
 #[derive(Debug, StructOpt)]
 pub enum AuthCommands {
@@ -18,12 +21,23 @@ pub enum AuthCommands {
     },
 }
 
+fn process_perms(perm: String) -> Result<Vec<String>, JsonRpcError> {
+    match perm.as_str() {
+        "admin" => Ok(ADMIN.to_owned()),
+        "sign" => Ok(SIGN.to_owned()),
+        "write" => Ok(WRITE.to_owned()),
+        "read" => Ok(READ.to_owned()),
+        _ => Err(JsonRpcError::INVALID_PARAMS),
+    }
+}
+
 impl AuthCommands {
     pub async fn run(&self) {
         match self {
             Self::CreateToken { perm } => {
                 let perm: String = perm.parse().unwrap();
-                print_rpc_res(auth_new(perm).await);
+                let perms = process_perms(perm).map_err(handle_rpc_err).unwrap();
+                print_rpc_res_bytes(auth_new((perms,)).await);
             }
         }
     }
