+++ forest/src/subcommand.rs
@@ -15,8 +15,10 @@ pub(super) async fn process(command: Subcommand) {
         Subcommand::Auth(cmd) => {
             cmd.run().await;
         }
-
         Subcommand::Genesis(cmd) => {
+            cmd.run().await;
+        }
+        Subcommand::Net(cmd) => {
             cmd.run().await;
         }
         Subcommand::Wallet(cmd) => {
