+++ node/forest_libp2p/Cargo.toml
@@ -46,6 +46,7 @@ async-trait = "0.1"
 lazy_static = "1.4"
 futures_cbor_codec = "0.3.1"
 git-version = "0.3.4"
+multihash = { version = "0.13", default-features = false, features = ["std", "blake2b", "derive"] }
 
 [dev-dependencies]
 forest_address = "0.3"
