+++ node/rpc-client/src/auth_ops.rs
@@ -1,23 +1,7 @@
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
+use crate::call_params;
+use jsonrpc_v2::Error;
+use rpc_api::auth_api::*;
 
-use super::client::filecoin_rpc;
-use auth::*;
-use jsonrpc_v2::Error as JsonRpcError;
-
-/// Creates a new JWT Token
-pub async fn auth_new(perm: String) -> Result<String, JsonRpcError> {
-    let perms = match perm.as_str() {
-        "admin" => ADMIN.to_owned(),
-        "sign" => SIGN.to_owned(),
-        "write" => WRITE.to_owned(),
-        "read" => READ.to_owned(),
-        _ => {
-            return Err(JsonRpcError::INVALID_PARAMS);
-        }
-    };
-
-    let ret: Vec<u8> = filecoin_rpc::auth_new((perms,)).await?;
-
-    Ok(String::from_utf8(ret)?)
+pub async fn auth_new(perm: AuthNewParams) -> Result<AuthNewResult, Error> {
+    call_params(AUTH_NEW, perm).await
 }
