+++ node/rpc-client/src/net_ops.rs
@@ -1,0 +1,19 @@
+use crate::call_params;
+use jsonrpc_v2::Error;
+use rpc_api::net_api::*;
+
+pub async fn net_addrs_listen(params: NetAddrsListenParams) -> Result<NetAddrsListenResult, Error> {
+    call_params(NET_ADDRS_LISTEN, params).await
+}
+
+pub async fn net_peers(params: NetPeersParams) -> Result<NetPeersResult, Error> {
+    call_params(NET_PEERS, params).await
+}
+
+pub async fn net_connect(params: NetConnectParams) -> Result<NetConnectResult, Error> {
+    call_params(NET_CONNECT, params).await
+}
+
+pub async fn net_disconnect(params: NetDisconnectParams) -> Result<NetDisconnectResult, Error> {
+    call_params(NET_DISCONNECT, params).await
+}
